implement SongPlayback

+++ src/api/java/work/lclpnet/kibu/nbs/data/Song.java
@@ -2,7 +2,7 @@ package work.lclpnet.kibu.nbs.data;
 
 import java.util.Map;
 
-public record Song(short durationTicks, float ticksPerSecond, SongMeta metaData, LoopConfig loopConfig,
+public record Song(int durationTicks, float ticksPerSecond, SongMeta metaData, LoopConfig loopConfig,
                    Map<Integer, Layer> layers, Instruments instruments) {
 
     public float durationSeconds() {
